Render structure renew:
    - textured (base class of all graphical objects that need a texture)
    - sprite : textured (Static rendering)
    - streamer : textured (Streaming access rendering)
    - layer : textured (Target access rendering)

That structure is based on the wsdl2::texture::access enumeration, in order
to simplify texture loading and access and make it safer.

Old layer class now it's called rendergroup

+++ engine/flatland.cpp
@@ -178,7 +178,7 @@ int flat::loop(float _fps)
     {
         do {
 
-            flatland_dt = 1.0f / fps + delay / CLOCKS_PER_SEC;
+            flatland_dt = 1.0f / fps + static_cast<float>(delay) / CLOCKS_PER_SEC;
 
             delay = clock();
